Extract Environments
